sonarparticlefilter fixed error with integer out of bounds sonar doesnt start timer, if there is no source
git-svn-id: svn+ssh://elmi.iti.uni-luebeck.de/export/svn/Hanse/branches/threadv1@884 8f13a6d9-6203-4c43-9b9f-a3039379152f

+++ Hanse/Module_SonarLocalization/sonarparticlefilter.cpp
@@ -383,7 +383,7 @@ void SonarParticleFilter::updateParticleFilter(const QList<QVector2D>& observati
     // resample
     QVector<QVector4D> resampledParticles(N);
     for (int i=0; i<N; i++) {
-        double r = 1.0*qrand()/(RAND_MAX + 1);
+        double r = 1.0*qrand()/((double)RAND_MAX + 1.0);
         int particle = 0;
         while (cumsum[particle] < r)
             particle++;
